requirements.txt
httpx
fastapi
uvicorn
rich
python-dotenv
notion-client
op
objprint

google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib